+++ packages/plugin-button/web/src/toolbar/buttonInputModal.jsx
@@ -8,6 +8,7 @@ import {
   Tab,
   FocusManager,
   BUTTON_SIZE,
+  SettingsMobileHeader,
 } from 'wix-rich-content-ui-components';
 import { KEYS_CHARCODE } from 'wix-rich-content-editor-common';
 import {
@@ -18,7 +19,6 @@ import {
 } from 'wix-rich-content-common';
 import DesignComponent from '../components/design-component';
 import SettingsComponent from '../components/settings-component';
-import Navbar from '../components/navbar';
 import PreviewComponent from '../components/preview-component';
 import { settingsTabValue, designTabValue } from '../constants';
 import styles from '../../statics/styles/button-input-modal.scss';
@@ -208,7 +208,12 @@ export default class ButtonInputModal extends Component {
     if (isMobile) {
       mobileView = (
         <div>
-          <Navbar onConfirm={this.onConfirm} onCancel={this.onCloseRequested} {...this.props} />
+          <SettingsMobileHeader
+            onSave={this.onConfirm}
+            onCancel={this.onCloseRequested}
+            theme={styles}
+            t={t}
+          />
           <PreviewComponent buttonObj={this.state} {...this.props} />
           <div className={styles.button_inputModal_scroll} ref={this.setScrollbarRef}>
             <div className={styles.button_inputModal_container} data-hook="ButtonInputModal">